adding a new test to show the setValue function
