+++ test/Server.test.js
@@ -1,7 +1,7 @@
 const expect = require("expect");
 const TreeServer = require("../server");
 const DeviceTree = require("../").DeviceTree;
-
+const {jsonRoot} = require("./utils");
 
 const LOCALHOST = "127.0.0.1";
 const PORT = 9009;
@@ -11,95 +11,13 @@ const wait = function(t) {
         setTimeout(resolve, t);
     });
 }
-const init = function(_src,_tgt) {
-    const targets = _tgt === undefined ? [ "tgt1", "tgt2", "tgt3" ] : _tgt;
-    const sources = _src === undefined ? [ "src1", "src2", "src3" ] : _src;
-    const labels = function(endpoints) {
-        let labels = [];
-        for (let i = 0; i < endpoints.length; i++) {
-            let endpoint = endpoints[i];
-            let l = { identifier: `Label-${i}` };
-            if (endpoint) {
-                l.value = endpoint;
-            }
-            labels.push(l);
-        }
-        return labels;
-    };
 
-    const buildConnections = function(s, t) {
-        let connections = [];
-        for (let i = 0; i < t.length; i++) {
-            connections.push({target: `${i}`});
-        }
-        return connections;
-    };
-
-    return [
-        {
-            // path "0"
-            identifier: "scoreMaster",
-            children: [
-                {
-                    // path "0.0"
-                    identifier: "identity",
-                    children: [
-                        {identifier: "product", value: "S-CORE Master"},
-                        {identifier: "company", value: "EVS"},
-                        {identifier: "version", value: "1.2.0"},
-                        {identifier: "author", value: "g.dufour@evs.com"}
-                    ]
-                },
-                {
-                    // path "0.1"
-                    identifier: "router",
-                    children: [
-                        {
-                            // path 0.1.0
-                            identifier: "matrix",
-                            type: "oneToN",
-                            mode: "linear",
-                            targetCount: targets.length,
-                            sourceCount: sources.length,
-                            connections: buildConnections(sources, targets),
-                            labels: ["0.1.1000"]
-                        },
-                        {
-                            identifier: "labels",
-                            // path "0.1.1000"
-                            number: 1000,
-                            children: [
-                                {
-                                    identifier: "targets",
-                                    // Must be 1
-                                    number: 1,
-                                    children: labels(targets)
-                                },
-                                {
-                                    identifier: "sources",
-                                    // Must be 2
-                                    number: 2,
-                                    children: labels(sources)
-                                },
-                                {
-				    identifier: "group 1",
-                                    children: [ {identifier: "sdp A", value: "A"}, {identifier: "sdp B", value: "B"}]
-				}
-                            ]
-                        }
-                    ]
-                }
-            ]
-        }
-    ];
-
-}
 describe("server", function() {
 
     describe("JSONtoTree", function() {
         let jsonTree;
         beforeAll(function() {
-            jsonTree = init();
+            jsonTree = jsonRoot();
         });
         it("should generate an ember tree from json", function() {
             const root = TreeServer.JSONtoTree(jsonTree);
@@ -115,7 +33,7 @@ describe("server", function() {
     describe("Server - Client communication", function() {
         let server,client;
         beforeAll(function() {
-            jsonTree = init();
+            jsonTree = jsonRoot();
             const root = TreeServer.JSONtoTree(jsonTree);
             server = new TreeServer(LOCALHOST, PORT, root);
             //server._debug = true;
@@ -153,9 +71,28 @@ describe("server", function() {
                     // Issue #33 TreeServer.handleGetDirectory does not subscribe to child parameters
                     expect(server.subscribers["0.0.0"]).toBeDefined();
                     // Keepalive
-	            client.disconnect();
+	                return client.disconnect();
                 });
         });
+        it("should be able to modify a parameter", () => {
+            client = new DeviceTree(LOCALHOST, PORT);
+            //client._debug = true;
+            return Promise.resolve()
+                .then(() => client.connect())
+                .then(() => {
+                    return client.getDirectory();
+                })
+                .then(() => client.expand(client.root.elements[0]))
+                .then(() => {
+                    expect(server.tree.elements[0].children[0].children[1].contents.value).not.toBe("gdnet");
+                    return client.setValue(client.root.elements[0].children[0].children[1], "gdnet");
+                })
+                .then(() => {
+                    expect(server.tree.elements[0].children[0].children[1].contents.value).toBe("gdnet");          
+                    return client.disconnect();
+                });
+        });
+        
 	it("should be able to get child with getNodeByPath", function() {
 	  //server._debug = true;
           client = new DeviceTree(LOCALHOST, PORT);
